rename package org.aion.type -> org.aion.types
