+++ src/module-info.java
@@ -1,6 +1,6 @@
 module aion.vm.api {
     exports org.aion.vm.api.interfaces;
-    exports org.aion.type;
+    exports org.aion.types;
     exports org.aion.interfaces;
     exports org.aion.util;
