+++ src/org/aion/vm/api/interfaces/TransactionSideEffects.java
@@ -2,7 +2,7 @@ package org.aion.vm.api.interfaces;
 
 import java.util.Collection;
 import java.util.List;
-import org.aion.type.Address;
+import org.aion.types.Address;
 
 /**
  * The side-effects caused by executing some transaction.
